requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==26.0
pillow==12.1.1
pygame==2.6.1
